plays and lastPlays working

+++ app/models/play.py
@@ -4,7 +4,7 @@ class Play(db.Model):
     __tablename__='plays'
 
     id = db.Column(db.Integer, primary_key=True)
-    playCount = db.Column(db.Numeric, nullable=False)
+    playCount = db.Column(db.Integer, nullable=False)
     userId = db.Column(db.Integer, db.ForeignKey('users.id'), nullable = False)
     albumId = db.Column(db.Integer, db.ForeignKey('albums.id'), nullable = False)
